+++ sentry/src/test/java/io/sentry/SentryClientTest.kt
@@ -2125,6 +2125,37 @@ class SentryClientTest {
   }
 
   @Test
+  fun `tombstone is added to the envelope from the hint`() {
+    val sut = fixture.getSut()
+    val attachment = Attachment.fromTombstone(byteArrayOf())
+    val hint = Hint().also { it.tombstone = attachment }
+
+    sut.captureEvent(SentryEvent(), hint)
+
+    verify(fixture.transport)
+      .send(
+        check { envelope ->
+          val tombstone = envelope.items.last()
+          assertNotNull(tombstone) { assertEquals(attachment.filename, tombstone.header.fileName) }
+        },
+        anyOrNull(),
+      )
+  }
+
+  @Test
+  fun `tombstone is dropped from hint via before send`() {
+    fixture.sentryOptions.beforeSend = CustomBeforeSendCallback()
+    val sut = fixture.getSut()
+    val attachment = Attachment.fromTombstone(byteArrayOf())
+    val hint = Hint().also { it.tombstone = attachment }
+
+    sut.captureEvent(SentryEvent(), hint)
+
+    verify(fixture.transport)
+      .send(check { envelope -> assertEquals(1, envelope.items.count()) }, anyOrNull())
+  }
+
+  @Test
   fun `capturing an error updates session and sends event + session`() {
     val sut = fixture.getSut()
     val scope = givenScopeWithStartedSession()
@@ -3647,6 +3678,7 @@ class SentryClientTest {
       hint.screenshot = null
       hint.viewHierarchy = null
       hint.threadDump = null
+      hint.tombstone = null
       return event
     }
   }